--- search/src/mapwisefox/search/adapters/dsl.py
@@ -1,3 +1,0 @@
-
-
-
