+++ search/src/mapwisefox/search/dsl/_enum.py
@@ -1,0 +1,4 @@
+from enum import StrEnum
+
+
+class
