+++ search/src/mapwisefox/search/dsl/adapters/__init__.py
@@ -1,0 +1,4 @@
+from ._base import DSLAdapter
+
+
+__all__ = ["DSLAdapter"]
