+++ search/src/mapwisefox/search/query_builder/adapters/__init__.py
