+++ search/src/mapwisefox/search/query_builder/adapters/_flat_output.py
