+++ search/src/mapwisefox/search/query_builder/adapters/_science_direct.py
