+++ search/src/mapwisefox/search/query_builder/adapters/_xplore.py
